split telegram post views
