+++ components/views/Video.tsx
@@ -1,0 +1,11 @@
+const TGVideo = ({ u }) => (
+  <video
+    controls={u.type !== 'animation'}
+    autoPlay={u.type === 'animation'}
+    loop
+  >
+    <source src={u.file_path} />
+  </video>
+);
+
+export default TGVideo;
